Write bottom-up solution

+++ oldbuild.cpp
@@ -4,8 +4,7 @@
 
 #include <vector>
 #include <algorithm>
-#include <numeric>
-#include "build.h"
+#include "build.hpp"
 using std::vector;
 
 bool containsCrossedBridges(const vector<Bridge> &subset)
@@ -18,7 +17,7 @@ bool containsCrossedBridges(const vector<Bridge> &subset)
     return subset[subset.size() - 2][0] >= subset.back()[0] || subset[subset.size() - 2][1] >= subset.back()[1];
 }
 
-int buildRecurse(const vector<Bridge> &bridges, vector<Bridge> &subset, int startIndex)
+int buildRecurse(const vector<Bridge> &bridges, vector<Bridge> &subset, size_t startIndex)
 {
     auto maxToll = 0;
     auto save = subset;
